requirements-worker.txt
# Transcription Engine
faster-whisper

# For Chinese Traditional/Simplified conversion
opencc-python-reimplemented
